Port to wheezy.

+++ SpecTcl/ScalerProcessorCallbacks.cpp
@@ -25,6 +25,7 @@
 #include <string>
 #include <buftypes.h>
 #include <buffer.h>
+#include <stdint.h>
 
 #include <histotypes.h>
 
@@ -214,7 +215,7 @@ CSpecTclControlBufferCallback::operator()(unsigned int nType, const void* pBuffe
 
   // All the other stuff comes from the buffer:
 
-  SetVariable(RunNumber, "%d", (void*)m_pProcessor->runNumber());
+  SetVariable(RunNumber, "%d", (void*)static_cast<uint64_t>(m_pProcessor->runNumber()));
   SetVariable(RunTitle,  "%s", (void*)m_pProcessor->Title().c_str());
